--- a/modules/givc/guivm.nix
+++ b/modules/givc/guivm.nix
@@ -31,6 +31,10 @@ in
         addr = hosts.${hostName}.ipv4;
         port = "9000";
       };
+      services = [
+        "reboot.target"
+        "poweroff.target"
+      ];
       admin = lib.head config.ghaf.givc.adminConfig.addresses;
       tls.enable = config.ghaf.givc.enableTls;
       enableUserTlsAccess = true;