--- a/modules/microvm/sysvms/guivm.nix
+++ b/modules/microvm/sysvms/guivm.nix
@@ -322,7 +322,7 @@ in
 
   config = lib.mkIf cfg.enable {
     microvm.vms."${vmName}" = {
-      autostart = true;
+      autostart = !config.ghaf.microvm-boot.enable;
       inherit (inputs) nixpkgs;
       config = guivmBaseConfiguration // {
         boot.kernelPackages =