<!--
    Copyright 2022-2025 TII (SSRC) and the Ghaf contributors
    SPDX-License-Identifier: CC-BY-SA-4.0
-->

## Description of Changes

Add custom microvm boot-order for ghaf to delay appvms startup to reserve resources for system-vms during boot.
This custom boot sequence is only enabled iff the gui-vm is enabled.

Features:
- remove the previous sleep for appvms by adding a wait-unit command to monitor the status of a service or target within a VM
- disable microvms autostart feature for net-vm, gui-vm, audio-vm, and app-vms, if enabled
- disable automatic restart of microvm@<vm> service. This allows to shutdown a machine internally (e.g., poweroff.target)

### Boot Targets

Appvms now have a boot priority flag that determines when they start booting; high, medium, low. Default is **medium**, with the following boot triggers:

| Trigger | Default | Appvm Priority | 
| ------- | ------- | ------- | 
| local-fs.target (host) | System VMs | high |
| greetd.target (guivm) | App VMs | medium |
| loginTarget** (guivm) | other | low |

** `ghaf-session.target` for labwc, `xdg-desktop-portal.service` for cosmic (slowest service as of now)

### Notes

**This patch does not aim to improve boot time or resource usage.** While minor improvements are noticeable, the main objective is to re-order the boot flow to allow system-vms to boot without significant appvm impact. In turn, the gui-vm should display the login-screen faster. The full boot from start menu to opening an application, without delay at the login screen, should pretty much match the previous (or be slightly faster).

Example boot CPU/IO profile without this boot order
![image](https://github.com/user-attachments/assets/c198ec64-c143-4110-b9d5-d86453252cd0)

Example boot CPU/IO  profile with this change
![image](https://github.com/user-attachments/assets/f4198bb8-c656-48ad-940d-a942be5f08cf)

Example measurements

Boot-to-Login: Time from starting the boot entry to the password/login screen showing
Login-to-Desktop: Time from login screen to desktop (top bar loaded)
Desktop-to-Chrome: Time from loaded desktop to started chrome application

|  | Boot-to-Login | Login-to-Desktop |  Desktop-to-Chrome | Total |
| ------- | ------- | ------- | ------- | ------- |
| Before Patch | 43 | 17 | 8 | 68 |
| | 44 | 16 | 8 | 68 |
| | 45 | 18 | 8 | 71 |
|  |  |  |  |  |
| After Patch |  31 | 22 | 12 | 65 |
| | 30 | 20 | 14 | 64 |
| | 31 | 18 | 14 | 63 |

If there is any user delay during login, the subsequent timing will be shorter.

### Type of Change
- [x] New Feature
- [ ] Bug Fix
- [x] Improvement / Refactor

## Related Issues / Tickets

<!--
Link to GitHub issues or JIRA tickets (if any) that this PR addresses or is related to
-->

## Checklist

<!--
Please check [X] for all items that apply. Leave [ ] if an item does not apply, but you have considered it.
Note that none of these are strict requirements — they are intended to inform reviewers.
Completing this checklist shows that you value and respect their time and effort.
-->

- [x] Clear summary in PR description
- [x] Detailed and meaningful commit message(s)
- [x] Commits are logically organized and squashed if appropriate
- [ ] [Contribution guidelines](https://github.com/tiiuae/ghaf/blob/main/CONTRIBUTING.md) followed
- [ ] Ghaf documentation updated with the commit - https://tiiuae.github.io/ghaf/
- [ ] Author has run `make-checks` and it passes
- [x] All automatic GitHub Action checks pass - see [actions](https://github.com/tiiuae/ghaf/actions)
- [x] Author has added reviewers and removed PR draft status

<!-- Additional description of omitted [ ] items if not obvious. -->

## Testing Instructions

### Applicable Targets
- [ ] Orin AGX `aarch64`
- [ ] Orin NX `aarch64`
- [x] Lenovo X1 `x86_64`
- [x] Dell Latitude `x86_64`

### Installation Method
- [ ] Requires full re-installation
- [x] Can be updated with `nixos-rebuild ... switch`
- [ ] Other: 

### Test Steps To Verify:
<!--
Provide clear, simple step-by-step instructions to verify the functionality.
Please do not assume that readers know everything you currently know.
-->
1. Verify that all VMs boot successfully with patch
2. Verify that all functionality works as with mainline
3. Compare boot with previous. Recommend to comment out the enable in `microvm-host.nix`, to still get systemd-bootcharts
```
microvm-boot = {
  #inherit (config.ghaf.virtualization.microvm.guivm) enable;
  debug = config.ghaf.profiles.debug.enable;
};
```
4. (Optional) analyze systemd bootchart
To copy a bootchart: `scp -J root@<ip of ghaf machine> root@ghaf-host:/run/log/*.svg .`