@vunnyso 2025-05-16
Hi @mbssrc this PR is nice improvement in boot time. Do we have plans to integrate it soon?