@mbssrc 2025-05-16
Problem with it is that although the UI shows up earlier, if one logs in without delay it takes equivalently longer to load the UI. Will look into it again soon