@milva-unikie 2025-05-30
Pre-merge-pipeline tests are failing for Lenovo-X1. The reason is degraded `systemctl status` in ghaf-host.
> Systemctl is not running! Status is degraded. Failed processes are: ['wait-for-login.service', 'wait-for-ui.service']